+++ orm-patterns-instances/active-jdbc/src/main/java/org/orm/patterns/instances/active/jdbc/service/PhoneServiceExecution.java
@@ -154,6 +154,9 @@ public class PhoneServiceExecution implements ApplicationListener<ApplicationRea
         }
     }
 
+    /**
+     * Delete execution.
+     */
     private void deleteExecution() {
         try {
             Collection<PhoneDetail> findPhoneResponse = new ArrayList<>();