+++ orm-patterns-instances/CHANGELOG.md
@@ -8,10 +8,17 @@ All notable changes to this project will be documented in this file.
 
 - Examples for Hibernate/spring data JPA.
 
+## 2020-11-10
+
+### Added
+- The find method for the phones to hibernate-jpa examples project.
+- The delete method for the phones to hibernate-jpa examples project.
+- The find service execution method to hibernate-jpa examples project.
+
 ## 2020-11-09
 
 ### Added
-- The update method for phone in hibernate-jpa examples project.
+- The update method for the phones to hibernate-jpa examples project.
 
 ## 2020-11-08